Updated install/uninstall scripts of cutncopy plugin

+++ plugins/cutncopy/uninstall.php
@@ -1,3 +1,6 @@
 <?php
-    bloecks_cutncopy_backend::deleteCookie();
+    if(class_exists('bloecks_cutncopy_backend'))
+    {
+        bloecks_cutncopy_backend::deleteCookie();
+    }
 ?>
